Wi-Fi MQTT/SOCK fix and AT Client fix
Wi-Fi: the devHandle field inside the local instance structure needed to be re-populated with the passed-in devHandle on initialisation.
AT Client: the change that introduced AT client hysteresis deleted what should be a pointless 10 ms delay for the EDM case.  However, it seems that, at least on some platforms (NINA-B1), removing this delay leads to timing issue on the AT interface, hence it is now reinstated.