Test fix for Windows: another attempt. (#579)
Commit 9af82ea52713bf212d275c0e7440be2167767598 and two that preceded it attempted to make the porting tests run reliably on the Windows platform, Windows appearing to be quite slow to start a thread when heavily loaded. Things have improved but there are still occasions when a thread isn't instantiated in time to lock a mutex within 10 seconds.

This commit introduces a flag in the started thread which is only set when the mutex has been locked, just to be completely sure that it has done its thing, with a guard time of 60 seconds. That should be enough; this code was run 10 times, all instances enabled to ensure maximum load on the Windows machine running the tests, and not a single portOs() failure was seen.