BREAKING CHANGE TO UART PORT API and fix in AT Client and Short Range code. (#583)
In the AT Client the uPortUartEventSend() function is used to "manually" retrigger URC handling after an AT command has been processed (i.e. when the AT stream has been unlocked with uAtClientUnlock()) if the UART buffer still contains some data, in case the received-but-unprocessed data contains a URC. However, the uPortUartEventSend() function blocks if the UART event queue is full, and it is called from within the AT client unlock function, which locks the AT client API; that will potentially prevent the elements of a URC from being processed as the AT client functions that are called to do so also lock the AT client API. All of this could lead to the AT client ending up in a mutex lock.

To fix this, uPortUartEventTrySend() is introduced, which will return if there is no room in the UART queue within a given time. This function is now used in the AT Client, and also in the Short Range code (where EDM does a similar thing), falling back to the uPortUartEventSend() where uPortUartEventTrySend() is not supported (which is the case on the Windows platform). Note that if uPortUartEventTrySend() fails to send because the queue is full that means there is already stuff in the event queue to trigger the URC handler and so this does not present a problem to the design.

IMPORTANT: if you do not wish to implement uPortUartEventTrySend(), you should have it return U_ERROR_COMMON_NOT_IMPLEMENTED and the code will continue to use the blocking call.