Location/GNSS API change: GNSS as a network on a cellular device. (#585)
With the new device/network split it is now possible to separate the GNSS device type from the GNSS network type, which means that we can use the same GNSS network type with a cellular device; this simplifies the customer-side of the code since there is no longer any need to copy handles around, the ubxlib code knows, internally, how to bring up a GNSS network on a cellular device.

Hence, with this commit, structures are modified as follows:

uDeviceGnssCfg_t:
- remove all of the "AT" items (devHandleAt is no longer required at all, gnssAtPinPwr and gnssAtPinDataReady move to uNetworkCfgGnss_t) and remove the somewhat duplicate GNSS transportType field,
- add a field includeNmea to perform the function of the GNSS transportType field, which is to indicate whether the UART output from the GNSS chip should include NMEA protocol data or just UBX protocol data,
- add a field pinDataReady, purely for forwards-compatibility; it is currently ignored,
- pinGnssEnablePower is renamed to pinEnablePower for consistency.

uNetworkCfgGnss_t:
- add devicePinPwr and devicePinDataReady (i.e. the old gnssAtPinPwr and gnssAtPinDataReady fields),
- add moduleType, since it can no longer rely on using the moduleType field in the device structure (which may now be the cellular module type),

uLocationAssist_t:
- remove devHandleAssist as it is also no longer required.

Functionally, to bring up a GNSS device that is either inside or attached via a cellular module, you do not call uDeviceOpen() on GNSS; instead you call uNetworkInterfaceUp() on the already-opened cellular device handle with the network type U_NETWORK_TYPE_GNSS and the configuration structure uNetworkCfgGnss_t. Once the GNSS network is up it can be configured, if required, with calls to the gnss API, as before, using the cellular device handle.

The operation of GNSS devices attached directly to your MCU is unchanged: call uDeviceOpen() and then uNetworkInterfaceUp() as normal.

This commit also changes the test code for the common network/location/sock/security and mqttClient tests so that it works properly with the new device/network API: the old gUNetworkTest[] array is now internal to u_network_shared_cfg.c and, instead of using that array directly, the tests request a list of device/network combinations to be tested by calling pUNetworkTestListAlloc(). The old U_NETWORK_TEST_TYPE_HAS_XXX() macros are mostly replaced by a set of callback functions which return true/false to indicate whether an entry in gUNetworkTest[] supports the given feature; for instance U_NETWORK_TEST_TYPE_HAS_SOCK() becomes uNetworkTestHasSock(). A test passes the appropriate callback function to pUNetworkTestListAlloc() so that the test is given only the device/network combinations that are right for it. Device handles are still stored centrally in gUNetworkTest[] (each device/network combination just gets a pointer to the relevant device handle); in this way it is possible to do a clean-up when required and there is no duplication of device handles.