Test change for short range only: default setting for the "reset to defaults" pin.
Since the short-range "reset to defaults" pin defined for a test instance is permanently connected to the DSR pin of the short-range module, the pin needs to be set to 0 (asserted) when it is not being toggled for resetting purposes or the module will get upset.  This commit introduces a preamble to the test sequence for short-range which does that.