AT client change only: perform processAsync() check in URC handler.
When the uAtClientIgnoreAsync() function was introduce the check that comes with it was not added to the functions in the URC handler, since that was already covered by locking the urcPermittedMutex during the AT client shut-down process.  However, when a short range device is shutting down it has to close the EDM stream, which is called by the AT client URC handler, _before_ the AT client is shut-down, which can lead to a mutex lock-up if a URC arrives at the wrong moment.

The short range device shut-down process already calls uAtClientIgnoreAsync() to protect itself, hence the mutex lock-up can be avoided if the processAsync() protection is extended to the URC handler.

Test: *