Test change for security credentials test: adapt to V5 uConnectExpress.
The security credentials test checks the subject of our stored test credential, except in the uConnectExpress case where no subject is returned.  However, with v5 uConnectExpress a subject _is_ returned but it is different to what we expect: for our "ubxlib_test_cert" we expect the subject to be "ubxlib client" (at least, this is what the cellular flavour of the AT+USECMNG=3 AT command returns) but uConnectExpress V5 has the subject as "CN=ubxlib ca", which seems to be the issuer.  So don't check the subject any more.