Cellular MQTT change only: updated quotes-check for MQTT messages. (#594)
During the introduction of MQTT-SN the processing of an outgoing MQTT/MQTT-SN message was changed such that, if it is fully ASCII, but it also contains a double quotation mark ("), then it will be sent in hex mode and not as ASCII; this is because the MQTT-SN client on the module does not accept quotation marks in ASCII-format MQTT messages (it strips them out) and it was assumed that the MQTT client is the same and we'd just not noticed because all non-SARA-R41X modules support binary send anyway.

However, things are more complicated than that: it turns out that SARA-R41X modules are fine with double-quotes characters but will reject a "," or ";" character ONLY if that character appears within double quotation marks; outside quotation marks it is fine. This was discovered because the natural order of the "string of all printable ASCII characters" in an ASCII table includes the double quotation mark character at the start of the sequence of punctuation characters, leading to a failure to MQTT-send, whereas a customer who was sending well-formed JSON (where the "," characters would naturally fall outside the double quotation marks) experienced no such failure.

Since converting a message to hex halves the maximum MQTT message length, and since SARA-R41x modules have shorter maximum MQTT publish lengths anyway (just 256 bytes maximum MQTT message length for SARA-R412M-02B, at least in FW version "M0.11.01,A.02.17"), the check is now updated to handle the more complex case.