Fix to short-range test code: shortRangeResetToDefaultSettings() correction.
In the test shortRangeResetToDefaultSettings(), the function pUShortRangePrivateGetInstance() is called to obtain the short-range instance pointer and, a little later, uShortRangeSetBaudrate() is called.  However, uShortRangeSetBaudrate() internally calls uShortRangeClose() and then uShortRangeOpenUart(), creating a new short-range instance.  The call to pUShortRangePrivateGetInstance() needs to be moved later so that it is called after the instance has been re-created.  Also, the stored handles for EDM, AT and potentially (though not actually) UART will have changed and these should be re-stored, just in case they are used again later.

The function descriptions for uShortRangeSetBaudrate() and pUShortRangePrivateGetInstance() are updated to make the side-effect of uShortRangeSetBaudrate() clear.

Test: 12 15.1 17 19 23 24