GNSS-via-cellular fix only: retry GNSS power-on. (#600)
On rare occasions, if a GNSS chip is connected via a cellular module and that cellular module is currently talking to the GNSS chip, an attempt to power the GNSS chip on over the AT interface will result in the error response "+CME ERROR: Invalid operation with LOC running / GPS Busy". As a workaround, the power-on procedure for GNSS-over-AT is now retried U_GNSS_AT_POWER_ON_RETRIES (2) times (i.e. a total of 3 times max.) with an interval of U_GNSS_AT_POWER_ON_RETRY_INTERVAL_SECONDS (10) seconds.