Fix to BLE configuration: avoid conflicting values.
Some BLE configuration items are linked in pairs, where one value must be set greater than or less than another.  For the configuration item pairs that may be set through the ubxlib BLE API, this is handled for the connection interval min/max pair (AT+UBTLECFG parameters 4 & 5) but was not handled for the scan interval/window pair (AT+UBTLECFG parameters 9 & 10).  It is now.