New Network API function: network status callback.
A function, uNetworkSetStatusCallback() is added, in the Network API, which allows a callback to be registered to obtain the status of any given network.  Primary in this status is whether the network is up or down and advice is given as to how the user should react to status changes; the actions will be DIFFERENT depending on the underlying network type.

A test for this function, networkOutage(), is added in the network tests; it is only run on a single test instance (randomly chosen to be the Windows one, test instance 23) since it requires a very specific and relatively expensive hardware setup (e.g. including a screened box for the cellular EVK, since cellular is too good at clinging on to coverage).