Test automation change only: increase wait time after USB/power is reapplied. (#608)
The automated test system resets all devices by USB/powering them off and on again at the start of a test run. It allows 5 seconds for all devices to come back up but it is possible that is not long enough when there are so many devices to re-detect; in particular, when running just instance 17, the flashing of the device always fails, nrfjprog complaining that a device of that serial number is not present, whereas if one runs everything instance 17 is detected and flashed perfectly fine, possibly because the Windows delay of bringing up so many things means that the flashing process for instance 17 is delayed in any case.

With this commit the time allowed for devices to come back up is extended from 5 seconds to 10 seconds, in which case running just instance 17 is fine. It might be that this is at the root of our occasional flash download failures in general.