GNSS UTC time clarification/improvement.
When a module is freshly delivered and a call is made to uGnssInfoGetTimeUtc() it will likely fail: this is because, in order to figure out the UTC time, leap second information has to be available, which is only broadcast from the satellites every 12.5 minutes (or it may be obtained using the u-blox CellLocate service).  For uGnssInfoGetTimeUtc() to work a position fix has to have been requested and the receiver then left on for long enough to receive the leap second information, which is stored in non-volatile memory for future use.

This commit adds a note about this and also adds the ability to set the UTC time standard (though this should likely be left in the default "automatic" mode).