Arduino fixes only: NINA-W102 compatibility and sockets.ino compilation errors. (#626)
Commit 5575a9ba2082919dcbdd95bb177a760b7799fabf temporarily disabled test instance 22, which is Arduino/ESP32 executing on a NINA-W102 board.  This was because the board on the test system entered a boot-loop after being programmed, hence we thought the HW was broken and so it was disabled until it could be replaced.

However, it turns out that the HW was absolutely fine, instead what had happened was that on 6th July 2022 version 2.0.4 of the ESP32 platform code for Arduino, defined in https://github.com/espressif/arduino-esp32/blob/gh-pages/package_esp32_index.json, was released.  This new version introduced compatibility with NINA-W106, which has 4 Mbytes of flash and is now the default, versus the 2 Mbytes of flash in NINA-W102.

To maintain compatibility with NINA-W102 the directive:

--build-property build.partitions=minimal

...needed to be added to the arduino-cli command-line.

Note that this is also being fixed over in the Espressif Arduino Gtihub repo: https://github.com/espressif/arduino-esp32/pull/7008.

Also, sockets.ino hadn't been properly updated to the new uDevice API; Arduino CLI obviously isn't re-building something when it should  Anyway, the file is updated now.