Cellular disconnect workaround motivated by SARA-R412M-02B. (#627)
There is a corner case that has occurred on SARA-R412M-02B when operating on an NB1 network (262 01 1nce.net) with a roaming SIM where the +CEREG URC indicates that we are still registered with the network, even though all other indications are that we have disconnected, and hence disconnect fails.  Hence we now also query the attach status (AT+CGATT?) while disconnecting and allow that to override all of the other indications.