GNSS change only: power on/off fixes. (#631)
The code under uGnssPwrOff() used the UBX-RXM-PMREQ command to put the GNSS chip into back-up mode; however the command was coded incorrectly and so did not actually work.  This was found during introduction of I2C support and, unfortunately, in the I2C case this command would not work anyway since I2C is not, of itself, able to take the GNSS chip out of back-up mode once more (i.e. waggling SDA or SCL does not wake the GNSS chip up, a physical HW reset via the RESET_N pin would be required).

Hence, with this commit, the power on/off process is performed using the UBX-CFG-RST command instead, applying GNSS stop/start as the mechanism, and the old uGnssPwrOff() functionality (fixed so that it actually works) is moved to a new API uGnssPwrOffBackup().