Feature: add support for cellular LARA-R6 module. (#633)
This commit adds iniitial support for the new cellular LARA-R6 module. This module is different to the currently supported set of cellular modules since it has LTE (Cat 1) but does NOT have Cat-M1/NB1. It also supports EGPRS and UTRAN, so a tri-RAT device.

Note: while this commit has passed the entire set of tests there is no LARA-R6 instance in the test farm as we are awaiting hardware for that, should arrive soon.

Also note that UART power saving is not enabled, there are issues if one does that which we are still looking into, will switch on UART power saving for LARA-R6 as a separate step.