Zephyr/NCS fix for NCS 2.0 and crypto. (#634)
In NCS 2.0 the MBEDTLS modules are brought in by NRF_SECURITY, which leads to a KConfig conflict because UBXLIB also brings them in. With this commit UBXLIB doesn't bring them in if NRF_SECURITY is on.

This is verified to not have an adverse effect on the existing NCS 1.6.1 integration and the customer who had the issue has verified that it removes the KConfig conflict on NCS 2.0.