Cellular LARA-R6 fix only: correct "assumed RAT".
In commit eb0b6e865221b1c8db3d7ba745d6087ac527d159 an "assumed RAT" was added to some of the cellular network functions since LARA-R6 sometimes forgets to say what RAT it is registered on in the +CEREG response, something we really must know.  However, that wasn't done correctly: the RAT passed to setNetworkStatus() is actually the 3GPP RAT number, not uCellNetRat_t; the naming of the parameter is now modified to reflect this and the other place where the RAT is required and might go missing, in registerNetwork() is also addressed.