Fix to cellular power test only.
In the cellPwr() test, specifically in testPowerAliveVInt(), there is a 30 second wait before powering off the module if VINT is not visible. However, when adding UART power saving some time ago this was accidentally cut and pasted into two places, resulting in an unnecessary additional 30 second delay.  This was spotted when watching the power test run for LARA-R6; it is now fixed.