Test application change for ESP-IDF only: set initial value of U_CFG_APP_PIN_CELL_RESET.
THe HPG Solution board that has a LARA-R6 and which we will shortly use in the ubxlib test farm has a pin of its NINA-W1 (ESP32) MCU wired to the RESET_N pin of LARA-R6.  This means that we have to put RESET_N, or U_CFG_APP_PIN_CELL_RESET, into the correct state at application startup.  This is now done at the start of appTask() in u_main.c (only if U_CFG_APP_PIN_CELL_RESET is >= 0).