GNSS change only: support for I2C and add M9. (#632)
Now that I2C is supported in the port layer, this commit adds support for GNSS over an I2C transport. Since I2C is being tested with an M9 module, initial support for M9 modules is also added; this will be expanded in future (e.g. to include setval/getval support).

Note that in this commit the uDeviceCfgI2c_t structure is modified WITHOUT changing its version: the structure was missing a field for the I2C HW block but, since I2C was never previously supported, the structure would never have been used and hence it is considered OK to modify it without a version change.