Arduino fixes only: copy example into output, increase warnings, cosmetic EDM fix. (#638)
The sockets.ino example wasn't being copied into the Arduino directory structure by u_arduino.py: it is now. Also, the default arduino-cli compiler warnings level for the ESP-IDF platform were being used, which is a little lenient: all warnings are now enabled with the exception of missing field initialisers and printf() format warnings, which we allow in the test/example code as it improves readability.

u_short_range_edm.c is changed just to set an automatic variable to a default value; the GCC compiler used for ESP-IDF complained that the variables were not initialised. In fact they are set by a pass-through function and the return value of that function is checked, so a problem could never occur, but making the change enables us to turn all warnings on for Arduino, which is a good thing.