Arduino platform build change only: no longer pre-compiled. (#646)
The ubxlib library metadata declared the library pre-compiled to save time, as there seemed penchant on the part of the Arduino tools to recompile for other purposes (something to do with parsing the code) and this stopped it.  However, it's probably not the right thing to do since we should recompile when a new version of code lands, make-style, hence reverting to not precompiled.