Device API, Port (I2C) API and GNSS fixes/API additions: uPortI2cAdopt() and set I2C address. (#647)
An new API is added to the I2C port layer: uPortI2cAdopt().  This does exactly the same as uPortI2cOpen() except that it does not modify any of the underlying HW when opening the I2C port.  This may be used if you are already using the I2C port for other purposes (e.g. reading from a sensor) and would like ubxlib to use it as you have already configured it.  Calling uPortClose() for an "adopted" I2C port will also not touch the underlying HW.

The uDeviceCfgI2c_t structure is modified to add, at the end, two fields: an int32_t "clockHertz" field and a boolean flag "alreadyOpen".  If "clockHertz" is non-zero then the device layer will set that clock frequency on the I2C bus. If the "alreadyOpen" flag is set to true then the device layer, when asked to communicate with a device via I2C, will use uPortI2cAdopt() instead of uPortI2cOpen() (and will ignore "clockHertz", since that would affect the I2C HW).  By default the compiler will set both fields to 0 when unused and hence backwards compatibility is maintained without the need to increment the version number of the structure.  FTAOD, the device layer does not use any of the other uPortI2cxxx() configuration functions.

A pair of new functions, uGnssSetI2cAddress()/uGnssGetI2cAddress() are added to the GNSS API: these allow the I2C address that will be used to communicate with a given GNSS device by the MCU to be set and to be read.

The uDeviceCfgGnss_t structure is modified to add, at the end, a field i2cAddress.  This field may be populated with the I2C address of the GNSS device, in which case the device layer will call uGnssSetI2cAddress() to set the target I2C address for that device.  By default the compiler will set this field to 0 if unused, the contents will be ignored and U_GNSS_I2C_ADDRESS will be used instead, maintaining backwards compatibility without the need to increment the version number of the structure.

Three fixes are also made in GNSS:

1) when the GNSS chip has to be reset, more than the 2 second power-on time is
   required (5 seconds is now used),
2) it wasn't permitted to open two GNSS instances on the same I2C port (d'oh!);
   it is now,
3) in the nRF5SDK and STM32F4 I2C implementations the stored SDA/SDC pins were
   reversed, which would have meant that the uPortI2cCloseRecoverBus() would
   not work; now fixed.