Ring buffer utility API additions: multiple readers and forced add. (#648)
The ring buffer utility is enhanced with a set of functions which allow there to be more than one reader of a ring buffer.  To use these functions the ring buffer must be created by calling uRingBufferCreateWithReadHandle() instead of uRingBufferCreate(); this maintains backwards compatibility with the existing interface, the only modification being that uRingBufferCreate() now returns a result code (since it creates a mutex there is a small possibility that it might fail) and the size of the ring buffer is now one less than the size of the linear buffer, rather than the same length, since one byte is required to prevent pointer wrapping.  Oh, and the pointer provided to the read functions can now be NULL to throw the data away.  But that's it.

To use this new functionality, after the ring buffer is created with uRingBufferCreateWithReadHandle(), a task may call uRingBufferTakeReadHandle() to obtain a read handle and then use uRingBufferReadHandle() with that handle to read from the buffer instead of uRingBufferRead(); uRingBufferRead() continues to work unless uRingBufferSetReadRequiresHandle() is called with true, at which point ONLY the "handled" read functions will return any data.  A read handle can be given back to the pool with uRingBufferGiveReadHandle().

Also new is uRingBufferForceAdd(): provided the data fits within the overall size of the ring buffer, this will add the data EVEN IF there is insufficient room, moving the read pointers on (i.e. losing data) as required.

A comprehensive test of the new implementation is also added.

These features are required in order to use the ring buffer code for reading data that has been streamed from a GNSS device and must be offered to multiple tasks.