GNSS changes only: new user message API and uGnssCfgGet/SetProtocolOut().
A new uGnssMsg API is introduced, allowing you to send user-defined messages to and monitor responses from the GNSS chip: both blocking and multiple asynchronous receives are supported.  As part of this the internals of GNSS message receive are re-written to use a ring buffer with an expanded version of the common utility ring buffer code; this has the twin advantages of allowing multiple readers of the incoming stream and allowing the application to read messages of arbitrary length, not limited by buffer sizes inside ubxlib.

IMPORTANT: this API currently ONLY works if the GNSS chip is directly connected to your MCU, i.e. UART and I2C; it does not [yet] work where the GNSS chip is connected via an intermediate cellular module (since that mechanism uses AT+UGUBX today).  Once ubxlib supports use of CMUX on the cellular interface it should be possible to extend this API for that case.

IMPORTANT: a consequence of this change is that uGnssUtilUbxTransparentSendReceive() should now ONLY be used if your GNSS module is connected via an intermediate cellular module.  This is because, for the directly-connected I2C/UART case, that function continues to use the old non-ring-buffer mechanism, talking directly to the I2C/UART, and hence could read and discard message data that you might have wanted in the ring-buffer.

Also, a new pair of functions are added in the GNSS configuration API allowing the type of protocol message sent by the GNSS chip to be set and read.  This was necessary in order to force NMEA messages to be output, making a proper test of the new GNSS message API.