Developer assist only: astyle.py (#658)
Add a Python script, astyle.py, in the root of the ubxlib directory tree, which calls the AStyle command line to auto-correct all code in the ubxlib directory tree. Saves having to remember the command-line. AStyle 3.1 must be installed and on the path for this to work. Must be called from the ubxlib root directory; this way it won't go trying to AStyle-ify, recursively, all of the files from the directory you happen to be in at the time...