Change to test code for cellular file system only: delete file at start of test. (#661)
If a test run is aborted it is possible for the test file used in the cellular file system test to still be present when the next test is run, which is a problem because the cellular file system write operation appends data to a file that is already there.  Hence we now make sure the file is deleted before the file system tests begin.  This is done at the start, rather than in the clean-up test at the end, since, in the "success" case, the clean-up test would have to reactivate the module to do this which costs more time.