More device callbacks added (#667)
Callbacks needed also for init and deinit of the device api added.
Also added some minor changes for the Zephyr board-ubxlib integration.
Removed CONFIG_USERSPACE from default Zephyr configuration. This is done as it is not needed and increases ram usage. This change only affect applications using ubxlib as a Zephyr module.