Cellular C2C fix to remove compiler warning on Zephyr. (#665)
The cellular C2C code needs access to rand(), which is provided by all platforms except that, for some reason, Zephyr doesn't expose a prototype for this in the usual way through stdlib.h, leading to a compiler warning.  The fix for this is already there in the code, to include u_port_clib_platform_specific.h, but the C2C code wasn't doing this.  It is now.