Feature, sockets: set local port. (#668)
It is now possible to set the local port number that will be used when an IP socket is created. This is useful if, for instance, you need to maintain a DTLS connection across multiple IP connections; call uSockSetNextLocalPort() with the port number you wish to use before calling uSockCreate(). Once uSockCreate() has been called the local port reverts to being IP-stack-assigned once more.

Note that this is obviously not thread-safe; it is up to the application to make it so if desired. Also note that neither cellular nor Wi-Fi supports reading the local port back again, so whether this actually works or not we can't easily verify, all we know is that the modules accept the command.