GNSS messaging improvement only: better performance. (#682)
When a partial message is received from the GNSS chip, more stuff needing to arrive before it can be decoded, the state of the message decode is now cached so that the whole message doesn't have to be re-parsed when the remainder turns up.  The GNSS message test code is updated to include a threshold minimum number of NMEA messages that should be received during its run-time.