Cellular feature: FOTA support. (#684)
This commit adds initial support for FOTA of cellular modules, i.e. firmware update over the air of the FW _inside_ the cellular module.  This is rudimentary support only: reading the status of the FW update process which is actually driven entirely by u-blox servers; the MCU has no control over the process whatsoever.  Support is limited to the SARA-R5, LARA-R6, SARA-R422 and SARA-R41XM-02B-03 series modules.

Note that there is currently no way to regression test this API; it has only been tested manually during development and only on a SARA-R5 module (which offers the widest support).