Fix to static_size "platform" only: OS selection by GNU make. (#701)
The mechanism used by the Makefile in static_size to determine what OS it is running on works fine except when Git is on the path: when Git is on the path a version of sh.exe is also on the path which GNU make spots and decides to use as its shell.  This then messes-up if any file paths contain slashes the wrong way around or brackets (e.g. "C:\Program Files (x86)").

The fix for this is to do an "ifeq ($(OS),Windows_NT)" check and force the shell to be cmd should this be the case.