BREAKING GNSS CHANGE for Windows/Linux platforms: port is now USB by default. (#698)
The ubxlib code needs to knows which physical port on the GNSS chip it is connected to.  Usually this is obvious - UART or I2C - however when the ubxlib code is running on a PC, talking directly to a GNSS chip over a USB port, it is not clear exactly how the communication is terminated on the GNSS-chip side - UART via an FTDI chip, or USB native to the GNSS chip (or maybe even I2C via an FTDI chip).

Commit c42582ba49f461b67c42c5450901225a2d0f197c introduced a conditional compilation flag, U_CFG_GNSS_PORT_NUMBER, which permitted the physical port number on the GNSS-chip-side assumed by this code to be overridden.

This commit makes a further modification: when this code is running on Windows/Linux, it will assume that it is connected to the USB port of the GNSS chip, not the UART port as before.

IF YOU ARE RUNNING THIS CODE ON WINDOWS/LINUX, CONNECTING DIRECTLY TO A GNSS CHIP THROUGH A UART (or via an FTDI adapter to the UART port of the GNSS chip) you MUST now set the conditional compilation flag U_CFG_GNSS_PORT_NUMBER=1 for your build.

If you are talking to a GNSS chip that is inside/connected-via a cellular module, or you are not running your code on Windows/Linux, nothing changes.