GNSS fix to NMEA decode: stop throwing valid messages away. (#705)
The NMEA decode function, uGnssPrivateDecodeNmea(), used when parsing the stream from the GNSS chip, does a length-check on the NMEA message.  However, it was checking the wrong thing (the buffer length rather than the message length) and ended up throwing perfectly valid NMEA messages away.  Many thanks to mazgch for spotting this.