BREAKING CHANGE, GNSS only: add VALSET/VALGET/VALDEL support and NACK-awareness. (#683)
GNSS modules from M9 onwards support a generic mechanism for setting/getting/deleting configuration values - the VALSET/VALGET/VALDEL ubx messages.  This commit adds support for that configuration mechanism in the uGnssCfg API.  As part of this change the uGnssMsg API is now able to detect and report when a NACK is received for a ubx-format message that it has been asked to monitor: should this occur the error code U_GNSS_ERROR_NACK will be returned.

BREAKING CHANGE: as a consequence of the latter change the uGnssMsg API introduced in commit c42582ba49f461b67c42c5450901225a2d0f197c is modified: the `size` parameter in the uGnssMsgReceiveCallback_t callback function signature becomes a signed int32_t instead of an unsigned size_t type.  This is so that, when a NACK is received for a ubx-format message class/ID that is being monitored, the callback can be called with the int32_t parameter set to U_GNSS_ERROR_NACK (-1025).  If you have any such callback functions their signatures must be modified appropriately and, in your callback function, you should check that what is now the errorCodeOrLength parameter is non-negative before calling uGnssMsgReceiveCallbackRead().