Cellular MQTT change only: remove redundant/dangerous assignment. (#709)
In the cellular MQTT entryFunction() the contents of ppInstance was being assigned to something (a) without first checking if ppInstance was NULL and (b) irrelevantly, since *ppInstance is assigned later on in the same function.  Not that the function was ever being given a NULL pointer for that variable but, you know, get it right.