GNSS info API addition: uGnssInfoGetCommunicationStats(). (#704)
A function, uGnssInfoGetCommunicationStats(), is added to retrieve the state of communications as seen by the GNSS chip.  In particular, this enables transmit and receive buffer overruns to be monitored; the GNSS message test now does this.  The underlying ubx message required to get this information is only supported on x9 GNSS modules and later.

Also, the automation database is updated so that more GNSS-capable instances will be targeted for branch runs, as well as pull-request runs.