GNSS test change only: RRLP poll rate and NMEA comprehender. (#714)
The rate of polling for RRLP messages in the GNSS non-blocking message test was set to 2 seconds. However, with all of the NMEA messages flowing also, and with only a 9600 bits/s UART link to the GNSS chip in the worst case, that's too fast, the RRLP messages back up; 3 seconds works.

Also, when polling for long RRLP messages, the GNSS chip may drop NMEA messages to make room, hence with this commit the first run around the GNSS non-blocking message test remains a mixed RRLP/NMEA message run, and is doubled in length, but with the NMEA comprehender switched off, then the second run is NMEA-only and has the NMEA comprehender switched on. The test run with a position fix included in it is removed: it wasn't used and, anyway, polling for RRLP messages is a much more stressful test.