GNSS fixes to UART buffer length and to timing, improved reliability on Windows.
The GNSS buffer size of 256 bytes was fine when all we had was a 9600 bits/s UART but on Windows with a USB interface that is too small, leading to timeouts when we go into polling mode waiting for more data.  A buffer of 1024 bytes works nicely.

Also, some of the timings, while waiting for data to arrive and when running the asynchronous message received task, were too short, leading to needless polling; this spotted while debugging the above.  They are now adjusted to give everyone more time.

Test: 1 2 6.1 6.2 11.0 11.1 12 14 15.0.0 15.1 17 18 19 22 23 25 gnss.port