Network API addition for cellular: user abort. (#722)
A field pKeepGoingCallback is added to the network configuration for cellular, permitting the user to pass in their own keep-going callback function, therefore allowing the abortable connect and disconnect operations to be aborted directly by the user if required. If the user does NOT require this feature, nothing changes: the field in the cellular network configuration structure will be zeroed by the compiler (i.e. pKeepGoingCallback= NULL) and so is ignored, timeoutSeconds will be applied instead, as it always was.