Test automation change only: warn when a test outcome is lost, rather than error. (#727)
On Nordic nRF52 platforms (both the NRF5SDK and the Zephyr flavours) it is possible for the line giving the outcome of a test to be lost. There is no pattern to this, it just happens sometimes, doesn't appear to be related to logging load. Since it is (a) rare enough not to be a worry in terms of checking code quality and (b) frequent enough to give irritating false negatives, each one of which has to be checked, with this commit we flag a warning when this happens rather than an error.