BREAKING GNSS change to uGnssMsg API: new message parsing and ring-buffer enhancements.  (#725)
With this change parsing of messages from the GNSS device is improved, now being performed within the ring-buffer itself, losing a copy while dodging lots of mutex locks.  Also, and this is the BREAKING aspect of this change, all messages are now fully FCS checked, i.e. for the UBX-format case the user no longer has to call uGnssMsgIsGood() and hence this function is removed; if you were previously using the gnss_cfgval_preview or development branches, where this function existed, you may need to remove any calls you made to it.

Support for parsing RTCM-format messages is added to the uGnssMsg API, though note that this is not yet fully tested; tests will be along shortly.

The uGnssMsg API also now permits return of messages it does not understand, i.e. unknown protocols, and return of any messages (all + unknown).