Feature: SPARTN message utilities. (#715)
These SPARTN protocol message utilities can be used for validating SPARTN (https://www.spartnformat.org/) messages, e.g. as received from a u-blox L-band module, e.g. NEO-D9S.  There is no _requirement_ to do this, SPARTN messages can simply be forwarded to the GNSS module, where the module itself will decode them.

Note that the _body_ of the SPARTN message is not decrypted/authenticated or anything by this code, all of that is done inside the u-blox GNSS chip that the SPARTN message is forwarded to.