Cellular change only: MNO profile database. (#726)
Some MNO profiles can cause a cellular module to behave in ways that do not currently work with ubxlib. For instance, the Verizon Wireless (VZW) profile, number 3, causes a cellular module to return a "not allowed" error to the command AT+CGDCONT.

With this commit a kind of "MNO profile database" is introduced in u_cell_mno_db.c/.h; we will use this to capture the specific behaviours that need special handling. The MNO profile is read when the cellular module boots, stored, and can be checked against this database later for such incompatibilities so that the ubxlib code can work with them. This is not intended to be a public API, just a way of ubxlib changing its own internal behaviour, versus MNO profile, dynamically.